pyproject.toml

[tool.pyright]
reportMissingImports = false
awaitNotInAsync = false
reportAwaitNotInAsync = false
reportTopLevelAwait = false
reportUnusedVariable = false
reportUnusedImport = false
enableExperimentalFeatures = true
reportAsyncNotInAsyncFunction = false


[tool.poetry]
name = "mpyc-web"
version = "0.3.0"
description = "MPyC for Multiparty Computation in Python"
authors = ["Berry Schoenmakers <berry@win.tue.nl>"]
packages = [
    { include = "./mpycweb" },
    { include = "./mpyc", from = "./mpyc" },
    { include = "./demos/*.py", from = "./mpyc" }
]

    [tool.poetry.dependencies]
    python = "^3.11,<3.13"
    rich = "^13.5.3"
    numpy = "^1.25.2"
    gmpy2 = "^2.1.5"
    pyyaml = "^6.0.1"

    [tool.poetry.group.dev.dependencies]
    python = "^3.11,<3.13"
    # didcomm = "0.3.0"
    # peerdid = "0.4.0"
    # qrcode = "^7.3.1"
    # numpy = "^1.23.4"
    # gmpy2 = "^2.1.2"
    packaging = "^23.0.0"
    # uvloop = "^0.17.0"
    # pyodide-pack = "^2.1.2"
    # pyodide-pack = "^0.2.0"
    # pyodide-pack = "^0.2.0"
    pyodide-py = "^0.24.1"
    rich = "^13.5.3"

    mpyc = { path = "./mpyc/", develop = true }
    matplotlib = "^3.5.2"
    pandas = "1.5.3"
    scipy = "1.11.2"
    lifelines = "^0.27.8"

[build-system]
requires = ["poetry-core", "setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"
